\name{halle}
\alias{halle}
\docType{data}

\title{
Halle data set
}

\description{
Artificial data set containing a total of 1600 observations in two dimensions (in three groups with 100, 1000, 500 obs. from top to bottom). Central cluster is quadratically transformed leading to a U-shaped main part of the data.
}

\usage{halle}

\examples{
# plot(halle)
}

\keyword{datasets}
